TFLiteSegmentation/requirements.txt
numpy==2.2.2
Pillow==11.1.0
tensorflow==2.18.0
